attempt to add prev next button functionality
