+++ src/components/PostList.js
@@ -1,29 +1,12 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import Excerpts from '../components/Excerpts'
 
-const PostList = ({ client }) => {
-  const [posts, setPosts] = useState([])
-
-  useEffect(() => {
-    const fetchPosts = () =>
-      client.getEntries({
-        content_type: 'blogPost',
-        select:
-          'sys.id,fields.title,fields.publishDate,fields.author,fields.content,fields.image,fields.snippet',
-        order: '-fields.publishDate'
-      })
-    fetchPosts().then(response => {
-      setPosts(response.items)
-    })
-  }, [client])
-
-  return (
-    <div className="flex-col">
-      {posts.map(({ fields, sys: { id } }, i) => (
-        <Excerpts key={id} index={i} id={id} limit={53} {...fields} />
-      ))}
-    </div>
-  )
-}
+const PostList = ({ posts }) => (
+  <div className="flex-col">
+    {posts.map(({ fields, sys: { id } }, i) => (
+      <Excerpts key={id} index={i} id={id} limit={53} {...fields} />
+    ))}
+  </div>
+)
 
 export default PostList
